fix dtype
see also: https://github.com/healx/pykeen/pull/1#discussion_r949165466

+++ src/pykeen/evaluation/rank_based_evaluator.py
@@ -510,7 +510,7 @@ class MacroRankBasedEvaluator(RankBasedEvaluator):
         # calculate key frequency
         inverse, counts = np.unique(keys, axis=0, return_inverse=True, return_counts=True)[1:]
         # weight = inverse frequency
-        weights = np.reciprocal(counts)
+        weights = np.reciprocal(counts, dtype=float)
         # broadcast to samples
         return weights[inverse]
